Bearer provider should support proper OAuth 2.0
