+++ go.mod
@@ -21,6 +21,7 @@ require (
 	github.com/spf13/viper v1.15.0
 	go.uber.org/fx v1.19.2
 	go.uber.org/zap v1.24.0
+	golang.org/x/oauth2 v0.8.0
 	golang.org/x/sync v0.1.0
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
@@ -89,11 +90,12 @@ require (
 	go.uber.org/multierr v1.10.0 // indirect
 	golang.org/x/crypto v0.7.0 // indirect
 	golang.org/x/mod v0.10.0 // indirect
-	golang.org/x/net v0.9.0 // indirect
+	golang.org/x/net v0.10.0 // indirect
 	golang.org/x/sys v0.8.0 // indirect
 	golang.org/x/text v0.9.0 // indirect
 	golang.org/x/time v0.3.0 // indirect
 	golang.org/x/tools v0.8.0 // indirect
+	google.golang.org/appengine v1.6.7 // indirect
 	google.golang.org/protobuf v1.30.0 // indirect
 	gopkg.in/ini.v1 v1.67.0 // indirect
 )
